test(constructor): consolidate mocks into mocks_test.go
Collect all interface mocks from construct_test.go,
construct_resource_test.go, construct_source_test.go, and
construct_callbacks_test.go into a single mocks_test.go for a clearer
overview of the test doubles.
