+++ bindings/go/constructor/construct_resource_test.go
@@ -9,7 +9,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"sigs.k8s.io/yaml"
 
-	"ocm.software/open-component-model/bindings/go/blob"
 	constructorruntime "ocm.software/open-component-model/bindings/go/constructor/runtime"
 	constructorv1 "ocm.software/open-component-model/bindings/go/constructor/spec/v1"
 	credconfigv1 "ocm.software/open-component-model/bindings/go/credentials/spec/config/v1"
@@ -18,177 +17,6 @@ import (
 	v2 "ocm.software/open-component-model/bindings/go/descriptor/v2"
 	"ocm.software/open-component-model/bindings/go/runtime"
 )
-
-// mockInputMethod implements ResourceInputMethod for testing
-type mockInputMethod struct {
-	processedResource *descriptor.Resource
-	processedBlob     blob.ReadOnlyBlob
-	capturedCreds     runtime.Typed
-}
-
-func (m *mockInputMethod) GetInputMethodScheme() *runtime.Scheme {
-	return runtime.NewScheme()
-}
-
-func (m *mockInputMethod) GetResourceCredentialConsumerIdentity(ctx context.Context, resource *constructorruntime.Resource) (identity runtime.Identity, err error) {
-	id := runtime.Identity{}
-	id.SetType(runtime.NewVersionedType("mock", "v1"))
-	return id, nil
-}
-
-func (m *mockInputMethod) ProcessResource(ctx context.Context, resource *constructorruntime.Resource, creds runtime.Typed) (*ResourceInputMethodResult, error) {
-	m.capturedCreds = creds
-	if m.processedResource != nil {
-		return &ResourceInputMethodResult{
-			ProcessedResource: m.processedResource,
-		}, nil
-	}
-	if m.processedBlob != nil {
-		return &ResourceInputMethodResult{
-			ProcessedBlobData: m.processedBlob,
-		}, nil
-	}
-	return nil, nil
-}
-
-// mockInputMethodProvider implements ResourceInputMethodProvider for testing
-type mockInputMethodProvider struct {
-	methods map[runtime.Type]ResourceInputMethod
-}
-
-func (m *mockInputMethodProvider) GetResourceInputMethod(ctx context.Context, resource *constructorruntime.Resource) (ResourceInputMethod, error) {
-	if method, ok := m.methods[resource.Input.GetType()]; ok {
-		return method, nil
-	}
-	return nil, fmt.Errorf("no input method resolvable for input specification of type %s", resource.Input.GetType())
-}
-
-type mockResourceRepository struct {
-	downloadData blob.ReadOnlyBlob
-	fail         bool
-
-	identityErr error
-}
-
-func (m *mockResourceRepository) GetResourceCredentialConsumerIdentity(ctx context.Context, resource *constructorruntime.Resource) (identity runtime.Identity, err error) {
-	if m.identityErr != nil {
-		return nil, m.identityErr
-	}
-	identity = runtime.Identity{}
-	identity.SetType(runtime.NewVersionedType("mock", "v1"))
-	return identity, nil
-}
-
-func (m *mockResourceRepository) DownloadResource(ctx context.Context, resource *descriptor.Resource, credentials runtime.Typed) (blob.ReadOnlyBlob, error) {
-	if m.fail {
-		return nil, fmt.Errorf("simulated download failure")
-	}
-	return m.downloadData, nil
-}
-
-// mockOwnershipAwareResourceRepository combines [mockResourceRepository] with
-// [mockOwnershipAttacher] so it satisfies both [ResourceRepository] and
-// [repository.OwnershipAwareRepository].
-type mockOwnershipAwareResourceRepository struct {
-	*mockResourceRepository
-	*mockOwnershipAttacher
-}
-
-func newMockOwnershipAwareResourceRepository() *mockOwnershipAwareResourceRepository {
-	return &mockOwnershipAwareResourceRepository{
-		mockResourceRepository: &mockResourceRepository{},
-		mockOwnershipAttacher:  &mockOwnershipAttacher{},
-	}
-}
-
-// mockResourceRepositoryProvider implements ResourceRepositoryProvider for testing.
-// Set err to exercise the "could not resolve the resource repository" branch.
-type mockResourceRepositoryProvider struct {
-	repo ResourceRepository
-	err  error
-}
-
-func (m *mockResourceRepositoryProvider) GetResourceRepository(ctx context.Context, resource *constructorruntime.Resource) (ResourceRepository, error) {
-	return m.repo, m.err
-}
-
-// mockAccess implements runtime.Typed for testing
-type mockAccess struct {
-	Type        string `json:"type"`
-	MediaType   string `json:"mediaType"`
-	Reference   string `json:"reference"`
-	Description string `json:"description"`
-}
-
-func (m *mockAccess) GetType() runtime.Type {
-	return runtime.NewVersionedType("mock", "v1")
-}
-
-func (m *mockAccess) SetType(typ runtime.Type) {
-	// No-op for testing
-}
-
-func (m *mockAccess) DeepCopyTyped() runtime.Typed {
-	return &mockAccess{
-		Type:        m.Type,
-		MediaType:   m.MediaType,
-		Reference:   m.Reference,
-		Description: m.Description,
-	}
-}
-
-// mockDigestProcessor implements ResourceDigestProcessor for testing
-type mockDigestProcessor struct {
-	processedDigest *descriptor.Digest
-}
-
-func (m *mockDigestProcessor) GetResourceRepositoryScheme() *runtime.Scheme {
-	return runtime.NewScheme()
-}
-
-func (m *mockDigestProcessor) GetResourceDigestProcessorCredentialConsumerIdentity(ctx context.Context, resource *descriptor.Resource) (identity runtime.Identity, err error) {
-	identity = runtime.Identity{}
-	identity.SetType(runtime.NewVersionedType("mock", "v1"))
-	return identity, nil
-}
-
-func (m *mockDigestProcessor) ProcessResourceDigest(ctx context.Context, resource *descriptor.Resource, credentials runtime.Typed) (*descriptor.Resource, error) {
-	if m.processedDigest != nil {
-		resource.Digest = m.processedDigest
-	}
-	return resource, nil
-}
-
-// mockDigestProcessorProvider implements ResourceDigestProcessorProvider for testing
-type mockDigestProcessorProvider struct {
-	processor ResourceDigestProcessor
-}
-
-func (m *mockDigestProcessorProvider) GetDigestProcessor(ctx context.Context, resource *descriptor.Resource) (ResourceDigestProcessor, error) {
-	return m.processor, nil
-}
-
-// mockCredentialProvider implements CredentialProvider for testing
-type mockCredentialProvider struct {
-	called      map[string]int
-	credentials map[string]map[string]string
-	fail        bool
-}
-
-func (m *mockCredentialProvider) Resolve(ctx context.Context, identity runtime.Identity) (runtime.Typed, error) {
-	m.called[identity.GetType().String()]++
-	if m.fail {
-		return nil, fmt.Errorf("simulated credential resolution failure")
-	}
-	creds := m.credentials[identity.GetType().String()]
-	if creds == nil {
-		return nil, nil
-	}
-	return &credconfigv1.DirectCredentials{
-		Type:       runtime.NewVersionedType(credconfigv1.CredentialsType, credconfigv1.Version),
-		Properties: creds,
-	}, nil
-}
 
 // setupTestComponent creates a basic component constructor for testing
 func setupTestComponent(t *testing.T, resourceYAML string) *constructorruntime.ComponentConstructor {
